backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
bcrypt==4.1.3
redis==5.0.7
python-multipart==0.0.9
boto3==1.34.162
psutil==6.0.0
slowapi==0.1.7
orjson==3.10.7
